requirements.txt
entmax==1.3
invoke==2.2.0
matplotlib==3.8.3
numpy==2.2.6
pandas==2.2.3
pyro_ppl==1.9.1
scikit_learn==1.4.1.post1
scipy==1.15.3
seaborn==0.13.2
torch==2.6.0
tqdm==4.65.0
wandb==0.19.11
